% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_rosa_atlas.R
\name{buildRosaAtlas}
\alias{buildRosaAtlas}
\title{Build Rosa chinensis Expression Atlas}
\usage{
buildRosaAtlas(
  counts_dir,
  metadata_file,
  strandedness = "reverse",
  min_samples = 10
)
}
\arguments{
\item{counts_dir}{Path to directory containing ReadsPerGene.out.tab files}

\item{metadata_file}{Path to SraRunTable.txt}

\item{strandedness}{One of "reverse", "forward", "unstranded"}

\item{min_samples}{Minimum samples a gene must be expressed in}
}
\value{
A SummarizedExperiment object
}
\description{
Build Rosa chinensis Expression Atlas
}
\examples{
\dontrun{
se <- buildRosaAtlas(
  counts_dir    = "path/to/counts",
  metadata_file = "path/to/SraRunTable.txt"
)
}
}
