% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importRosaCounts}
\alias{importRosaCounts}
\title{Import STAR count files into OrnAtlas}
\usage{
importRosaCounts(file_paths, col_data, strandedness = "reverse")
}
\arguments{
\item{file_paths}{Named character vector of paths to
ReadsPerGene.out.tab files. Names become sample IDs.}

\item{col_data}{A data.frame with sample metadata.}

\item{strandedness}{One of "reverse", "forward", "unstranded"}
}
\value{
A SummarizedExperiment object
}
\description{
Import STAR count files into OrnAtlas
}
\examples{
\dontrun{
files <- c(Stage1 = "path/to/sample1/ReadsPerGene.out.tab")
importRosaCounts(files, col_data = metadata)
}
}
