% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{ornTheme}
\alias{ornTheme}
\title{OrnAtlas ggplot2 theme}
\usage{
ornTheme(base_size = 12)
}
\arguments{
\item{base_size}{Base font size}
}
\value{
A ggplot2 theme
}
\description{
OrnAtlas ggplot2 theme
}
