% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas.R
\name{plotAtlas}
\alias{plotAtlas}
\title{Plot expression atlas heatmap}
\usage{
plotAtlas(atlas, top_n = 50, title = "Expression Atlas")
}
\arguments{
\item{atlas}{A SummarizedExperiment from buildAtlas()}

\item{top_n}{Number of top variable genes to show}

\item{title}{Plot title}
}
\value{
A ggplot2 heatmap object
}
\description{
Plot expression atlas heatmap
}
\examples{
\donttest{
data(rosa_example)
se_norm <- normalizeCounts(rosa_example, method = "CPM")
atlas <- buildAtlas(se_norm, group_var = "tissue_simple",
                    assay_name = "CPM")
plotAtlas(atlas, top_n = 30, title = "Rosa chinensis Atlas")
}
}
