% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plotPCA}
\alias{plotPCA}
\title{PCA plot of expression data}
\usage{
plotPCA(
  se,
  assay_name = "counts",
  color_by = "stage",
  shape_by = NULL,
  label_points = TRUE
)
}
\arguments{
\item{se}{A SummarizedExperiment object}

\item{assay_name}{Which assay to use}

\item{color_by}{Column in colData for point color}

\item{shape_by}{Column in colData for point shape}

\item{label_points}{Show sample labels}
}
\value{
A ggplot2 PCA plot
}
\description{
PCA plot of expression data
}
\examples{
data(rosa_example)
se_norm <- normalizeCounts(rosa_example, method = "CPM")
}
