% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas.R
\name{plotAtlas}
\alias{plotAtlas}
\title{Plot expression atlas heatmap}
\usage{
plotAtlas(atlas, top_n = 50, title = "Expression Atlas")
}
\arguments{
\item{atlas}{A SummarizedExperiment from buildAtlas()}

\item{top_n}{Number of top variable genes to show}

\item{title}{Plot title}
}
\value{
A ggplot2 heatmap object
}
\description{
Plot expression atlas heatmap
}
\examples{
\dontrun{
plotAtlas(atlas, top_n = 50, title = "Rose Petal Atlas")
}
}
