% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_analysis.R
\name{runDE}
\alias{runDE}
\title{Differential Expression Analysis}
\usage{
runDE(
  se,
  design = ~tissue_clean,
  contrast = NULL,
  alpha = 0.05,
  lfc_threshold = 1
)
}
\arguments{
\item{se}{A SummarizedExperiment object}

\item{design}{Formula for experimental design}

\item{contrast}{Character vector of length 3}

\item{alpha}{Significance threshold}

\item{lfc_threshold}{Minimum log2 fold change}
}
\value{
A data.frame of DE results
}
\description{
Differential Expression Analysis
}
\examples{
\dontrun{
results <- runDE(se,
  design   = ~ tissue_clean,
  contrast = c("tissue_clean", "petal", "abscission zone"))
}
}
